requirements.txt
Flask>=1.0,<=1.1.2
pandas==1.3.2
requests==2.26.0
xlsxwriter==3.0.2
python-dotenv==0.19.2